changed package name to lowercase
