--- backend/src/main/java/com/htecgroup/SkyNest/SkyNestApplication.java
@@ -1,25 +1,0 @@
-package com.htecgroup.SkyNest;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-@SpringBootApplication
-public class SkyNestApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(SkyNestApplication.class, args);
-  }
-
-  @Bean
-  public BCryptPasswordEncoder bCryptPasswordEncoder() {
-    return new BCryptPasswordEncoder();
-  }
-
-  @Bean
-  public ModelMapper modelMapper() {
-    return new ModelMapper();
-  }
-}
