--- backend/src/main/java/com/htecgroup/SkyNest/model/dto/RoleDto.java
@@ -1,12 +1,0 @@
-package com.htecgroup.SkyNest.model.dto;
-
-import lombok.Data;
-
-import java.util.UUID;
-
-@Data
-public class RoleDto {
-
-  private UUID id;
-  private String name;
-}
