--- backend/src/main/java/com/htecgroup/SkyNest/model/request/UserLoginRequest.java
@@ -1,9 +1,0 @@
-package com.htecgroup.SkyNest.model.request;
-
-import lombok.Data;
-
-@Data
-public class UserLoginRequest {
-  private String email;
-  private String password;
-}
