--- backend/src/main/java/com/htecgroup/SkyNest/model/request/UserRegisterRequest.java
@@ -1,14 +1,0 @@
-package com.htecgroup.SkyNest.model.request;
-
-import lombok.Data;
-
-@Data
-public class UserRegisterRequest {
-
-  private String email;
-  private String password;
-  private String name;
-  private String surname;
-  private String phoneNumber;
-  private String address;
-}
