--- backend/src/main/java/com/htecgroup/SkyNest/model/response/ErrorMessage.java
@@ -1,12 +1,0 @@
-package com.htecgroup.SkyNest.model.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ErrorMessage {
-  private String message;
-  private int status;
-  private String timestamp;
-}
