--- backend/src/main/java/com/htecgroup/SkyNest/model/response/UserResponse.java
@@ -1,14 +1,0 @@
-package com.htecgroup.SkyNest.model.response;
-
-import lombok.Data;
-
-@Data
-public class UserResponse {
-
-  private String id;
-  private String email;
-  private String name;
-  private String surname;
-  private String phoneNumber;
-  private String address;
-}
