--- backend/src/main/java/com/htecgroup/SkyNest/repository/RoleRepository.java
@@ -1,13 +1,0 @@
-package com.htecgroup.SkyNest.repository;
-
-import com.htecgroup.SkyNest.model.enitity.RoleEntity;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface RoleRepository extends CrudRepository<RoleEntity, UUID> {
-  Optional<RoleEntity> findByName(String name);
-}
