--- backend/src/main/java/com/htecgroup/SkyNest/repository/UserRepository.java
@@ -1,15 +1,0 @@
-package com.htecgroup.SkyNest.repository;
-
-import com.htecgroup.SkyNest.model.enitity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, UUID> {
-  Optional<UserEntity> findUserByEmail(String email);
-
-  boolean existsByEmail(String email);
-}
