--- backend/src/main/java/com/htecgroup/SkyNest/security/WebSecurity.java
@@ -1,39 +1,0 @@
-package com.htecgroup.SkyNest.security;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class WebSecurity extends WebSecurityConfigurerAdapter {
-
-  private final CustomUserDetailsService userDetailsService;
-  private final BCryptPasswordEncoder bCryptPasswordEncoder;
-
-  @Override
-  protected void configure(HttpSecurity http) throws Exception {
-    http.cors()
-        .and()
-        .csrf()
-        .disable()
-        .authorizeRequests()
-        .antMatchers(HttpMethod.POST, "/users/register")
-        .permitAll()
-        //        .anyRequest()
-        //        .authenticated()
-        .and()
-        .sessionManagement()
-        .sessionCreationPolicy(SessionCreationPolicy.STATELESS);
-  }
-
-  @Override
-  protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-    auth.userDetailsService(userDetailsService).passwordEncoder(bCryptPasswordEncoder);
-  }
-}
