--- backend/src/main/java/com/htecgroup/SkyNest/service/UserService.java
@@ -1,10 +1,0 @@
-package com.htecgroup.SkyNest.service;
-
-import com.htecgroup.SkyNest.model.dto.UserDto;
-
-public interface UserService {
-
-  UserDto registerUser(UserDto userDto);
-
-  UserDto findUserByEmail(String email);
-}
