+++ backend/src/main/java/com/htecgroup/skynest/exception/UserException.java
@@ -1,0 +1,18 @@
+package com.htecgroup.skynest.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor
+public class UserException extends RuntimeException {
+
+  private String message;
+  private HttpStatus status;
+
+  public UserException(UserExceptionType userExceptionType) {
+    message = userExceptionType.getMessage();
+    status = userExceptionType.getStatus();
+  }
+}
