+++ backend/src/main/java/com/htecgroup/skynest/model/dto/RoleDto.java
@@ -1,0 +1,12 @@
+package com.htecgroup.skynest.model.dto;
+
+import lombok.Data;
+
+import java.util.UUID;
+
+@Data
+public class RoleDto {
+
+  private UUID id;
+  private String name;
+}
