+++ backend/src/main/java/com/htecgroup/skynest/model/request/UserLoginRequest.java
@@ -1,0 +1,9 @@
+package com.htecgroup.skynest.model.request;
+
+import lombok.Data;
+
+@Data
+public class UserLoginRequest {
+  private String email;
+  private String password;
+}
