+++ backend/src/main/java/com/htecgroup/skynest/model/request/UserRegisterRequest.java
@@ -1,0 +1,14 @@
+package com.htecgroup.skynest.model.request;
+
+import lombok.Data;
+
+@Data
+public class UserRegisterRequest {
+
+  private String email;
+  private String password;
+  private String name;
+  private String surname;
+  private String phoneNumber;
+  private String address;
+}
