+++ backend/src/main/java/com/htecgroup/skynest/model/response/ErrorMessage.java
@@ -1,0 +1,12 @@
+package com.htecgroup.skynest.model.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ErrorMessage {
+  private String message;
+  private int status;
+  private String timestamp;
+}
