+++ backend/src/main/java/com/htecgroup/skynest/service/UserService.java
@@ -1,0 +1,10 @@
+package com.htecgroup.skynest.service;
+
+import com.htecgroup.skynest.model.dto.UserDto;
+
+public interface UserService {
+
+  UserDto registerUser(UserDto userDto);
+
+  UserDto findUserByEmail(String email);
+}
