--- backend/src/test/java/com/htecgroup/SkyNest/SkyNestApplicationTests.java
@@ -1,13 +1,0 @@
-package com.htecgroup.SkyNest;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SkyNestApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
