+++ backend/src/test/java/com/htecgroup/skynest/SkyNestApplicationTests.java
@@ -1,0 +1,13 @@
+package com.htecgroup.skynest;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SkyNestApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
